+++ Code/ConfigOptions.cs
@@ -4,6 +4,7 @@ namespace ArtifactOfTheUnchainedMod
 {
     public static class ConfigOptions
     {
+        public static ConfigEntry<bool> EnableDebugLogging;
         public static ConfigEntry<bool> ArtifactlessMode;
         public static ConfigEntry<string> ItemProcNerfBodyBlacklist;
         internal static void ItemProcNerfBodyBlacklist_SettingChanged(object sender, System.EventArgs e)
@@ -11,6 +12,8 @@ namespace ArtifactOfTheUnchainedMod
             ItemProcNerfBodyBlacklistArray = ItemProcNerfBodyBlacklist.Value.Split(',');
         }
         internal static string[] ItemProcNerfBodyBlacklistArray;
+        public static ConfigEntry<bool> PreventAllItemChaining;
+        public static ConfigEntry<float> NerfProcDamagePerEthereal;
 
         public static ConfigEntry<float> ProcChainDamageNerfToPercent;
         public static ConfigEntry<float> ProcChainCoefficientNerfToPercent;
@@ -22,10 +25,13 @@ namespace ArtifactOfTheUnchainedMod
         public static ConfigEntry<float> ProcFromEquipmentDamageNerfToPercent;
         public static ConfigEntry<float> ProcFromEquipmentCoefficientNerfToPercent;
 
-        public static ConfigEntry<bool> PreventAllItemChaining;
 
         public static void BindConfigEntries(ConfigFile Config, ArtifactBase artifactInstance)
         {
+            EnableDebugLogging = Config.Bind<bool>(
+                "Misc",
+                "Enable debug logging", false
+            );
             ArtifactlessMode = Config.Bind<bool>(
                 "Misc",
                 "Run without requiring artifact", false,
@@ -47,12 +53,17 @@ namespace ArtifactOfTheUnchainedMod
                 "Prevent ALL items from proccing other items", false,
                 "Should not just procs from other procs but item procs from ANY item be blocked? This means that ONLY damage from skills can proc your items."
             );
+            NerfProcDamagePerEthereal = Config.Bind<float>(
+                "Misc",
+                "Nerf proc damage per ethereal level", 0.07f,
+                "ONLY WORKS IF YOU'RE RUNNING THE STARSTORM 2 BETA. Nerfs ALL proc damage based on how many ethereal teleporters you've completed. Because this applies to EVERY proc anything does, it will apply multiple times in a proc chain.\nFormula: 1 - (This config's value * # of Ethereal teleporters completed)"
+            );
 
 
             ProcChainDamageNerfToPercent = Config.Bind<float>(
                 "Procs from proc chains",
-                "Nerf damage", 0.1f,
-                "Should damage from chained procs (i.e. atg procced by a charged perforator hit) be reduced to a percent of what they would normally do? I.E. 0.2 is 20% of the normal damage, 0.02 is 2%, and 1 is vanilla behavior as there is no change."
+                "Nerf damage", 0.4f,
+                "Should damage from chained procs (i.e. atg procced by a charged perforator hit) be reduced to a percent of what they would normally do? I.E. 0.4 is 40% of the normal damage, 0.04 is 4%, and 1 is vanilla behavior as there is no change."
             );
             ProcChainCoefficientNerfToPercent = Config.Bind<float>(
                 "Procs from proc chains",
@@ -68,8 +79,8 @@ namespace ArtifactOfTheUnchainedMod
 
             ProcFromItemDamageNerfToPercent = Config.Bind<float>(
                 "Procs from other items",
-                "Nerf damage", 0.2f,
-                "Should damage for procs from items (i.e. fireworks or royal capacitor) be nerfed to a percent of what it would normally be? I.E. 0.2 is 20% of the normal damage, 0.02 is 2%, and 1 is vanilla behavior as there is no change."
+                "Nerf damage", 0.15f,
+                "Should damage for procs from items (i.e. fireworks or royal capacitor) be nerfed to a percent of what it would normally be? I.E. 0.15 is 15% of the normal damage, 0.05 is 5%, and 1 is vanilla behavior as there is no change."
             );
             ProcFromItemCoefficientNerfToPercent = Config.Bind<float>(
                 "Procs from other items",
@@ -80,7 +91,7 @@ namespace ArtifactOfTheUnchainedMod
 
             ProcFromEquipmentDamageNerfToPercent = Config.Bind<float>(
                 "Procs from equipments",
-                "Nerf damage", 0.2f,
+                "Nerf damage", 0.25f,
                 "Should damage for procs from equipments (i.e. sawmerang or royal capacitor) be nerfed to a percent of what it would normally be? I.E. 0.2 is 20% of the normal damage, 0.02 is 2%, and 1 is vanilla behavior as there is no change."
             );
             ProcFromEquipmentCoefficientNerfToPercent = Config.Bind<float>(