+++ Code/Log.cs
@@ -13,22 +13,15 @@ namespace ArtifactOfTheUnchainedMod
 
         internal static void Debug(object data)
         {
-#if DEBUG
-            _logSource.LogDebug(data);
-#endif
+            if (ConfigOptions.EnableDebugLogging.Value)
+            {
+                _logSource.LogDebug(data);
+            }
         }
         internal static void Info(object data) => _logSource.LogInfo(data);
         internal static void Message(object data) => _logSource.LogMessage(data);
         internal static void Warning(object data) => _logSource.LogWarning(data);
         internal static void Error(object data) => _logSource.LogError(data);
         internal static void Fatal(object data) => _logSource.LogFatal(data);
-        internal static void NerfedProc(string message)
-        {
-            if (!Main.AllowLoggingNerfs)
-            {
-                return;
-            }
-            Info(message);
-        }
     }
 }